pre-developed apiHandler -> views services
